Reorganize the test folder.
