+++ ava.config.mjs
@@ -5,7 +5,7 @@ import { fileURLToPath } from "url";
 // Use the node tsconfig import paths.
 const __filename = fileURLToPath(import.meta.url);
 const __dirname = path.dirname(__filename);
-process.env.TSX_TSCONFIG_PATH = path.join(__dirname, "test", "tsconfig.json");
+process.env.TSX_TSCONFIG_PATH = path.join(__dirname, "tsconfig.test.json");
 
 export default {
   extensions: {
