+++ test/fixtures/browser.test.ts.json
@@ -5,24 +5,26 @@
         "path": "/robots.txt",
         "body": "",
         "status": 200,
-        "response": "<!doctype html><html lang=\"en\"><head><meta charset=\"utf-8\"/><meta name=\"viewport\" content=\"width=device-width,initial-scale=1,shrink-to-fit=no\"/><meta name=\"mobile-web-app-capable\" content=\"yes\"/><meta name=\"format-detection\" content=\"telephone=no\"/><title>Sindri</title><link rel=\"preconnect\" href=\"https://fonts.googleapis.com\"/><link rel=\"preconnect\" href=\"https://fonts.gstatic.com\" crossorigin/><link href=\"https://fonts.googleapis.com/css2?family=Inter:wght@400;500;600;800&display=swap\" rel=\"stylesheet\"/><script defer=\"defer\" src=\"/runtime.30ef403476e3e0b7e5126f2bf73adaf4.js\"></script><script defer=\"defer\" src=\"/npm.mui.0b911dc429e4a0063355e6727bfda865.js\"></script><script defer=\"defer\" src=\"/npm.core-js.4a4d82d0a402a1542119c48bec590bc2.js\"></script><script defer=\"defer\" src=\"/npm.moment.7ade13edecb1a25b3c7970b6cb2f3988.js\"></script><script defer=\"defer\" src=\"/npm.react-redux.9a6d3de1ceff07c4a1e29b3371cd73ed.js\"></script><script defer=\"defer\" src=\"/npm.emotion.f7332e6603cd98c6094b6f8899fcfa6b.js\"></script><script defer=\"defer\" src=\"/npm.redux-persist.4c4b78a1b7c8c7f327d4aeb3a7c23445.js\"></script><script defer=\"defer\" src=\"/npm.react-router.6f25fa7680daa338de5b4b74b5f8a2c2.js\"></script><script defer=\"defer\" src=\"/npm.react-transition-group.d98ae7fc5c853d3e99debcd8d05ddc19.js\"></script><script defer=\"defer\" src=\"/npm.reduxjs.6c9b43ff2be88131d7e53cd019f54782.js\"></script><script defer=\"defer\" src=\"/npm.floating-ui.627b2f7cd13096c7a6803d4e65f58ae9.js\"></script><script defer=\"defer\" src=\"/npm.styled-components.717cac08aa7d011c525fee66b5846991.js\"></script><script defer=\"defer\" src=\"/npm.react-dropzone.6188accaaa08e13a91b1464693bde3aa.js\"></script><script defer=\"defer\" src=\"/npm.react-dom.689ade6574851e1962f847b3d5b8bfa5.js\"></script><script defer=\"defer\" src=\"/npm.sweetalert2.5f59dfd36c3ecbb9cd472c0e76133d6b.js\"></script><script defer=\"defer\" src=\"/npm.popperjs.4789ecd07935921773dffa5dfdbdbed5.js\"></script><script defer=\"defer\" src=\"/895.208c34bb5c3ecb0422a91ceaa91094d4.js\"></script><script defer=\"defer\" src=\"/main.6f83d425c82354878fbbaf1db76ab982.js\"></script><meta name=\"apple-mobile-web-app-title\" content=\"React Boilerplate\" /><meta name=\"apple-mobile-web-app-capable\" content=\"yes\" /><meta name=\"apple-mobile-web-app-status-bar-style\" content=\"default\" /><meta name=\"theme-color\" content=\"#b1624d\" /><link rel=\"apple-touch-icon\" sizes=\"180x180\" href=\"/icon_180x180.63c71ff1fe800fe7fc6a16584342cb4c.png\" /><link rel=\"apple-touch-icon\" sizes=\"167x167\" href=\"/icon_167x167.359d468d6c991c5c01aa3cdcfaeea8bd.png\" /><link rel=\"apple-touch-icon\" sizes=\"152x152\" href=\"/icon_152x152.3b0f40a9c71e9ad8aa01e085feaa48f7.png\" /><link rel=\"apple-touch-icon\" sizes=\"120x120\" href=\"/icon_120x120.f51c13266175c3794ce042c18f41966b.png\" /><link rel=\"manifest\" href=\"/manifest.8259ad0cff5bad38962209630528b19c.json\" /></head><body><noscript>If you're seeing this message, that means <strong>JavaScript has been disabled on your browser</strong>, please <strong>enable JS</strong> to make this app work.</noscript><div id=\"app\"></div><link href=\"https://cdn.jsdelivr.net/npm/@sweetalert2/theme-material-ui/material-ui.css\" rel=\"stylesheet\"/></body></html>",
+        "response": "#########################\n##   ______  ______    ##\n##  |__  / |/ /  _ \\   ##\n##    / /| ' /| |_) |  ##\n##   / /_| . \\|  __/   ##\n##  /____|_|\\_\\_|      ##\n##                     ##\n#########################\n\nUser-agent: *\nDisallow: /admin/\nDisallow: /api/\n\nSitemap: https://sindri.app/docs/sitemap.xml\n",
         "rawHeaders": [
-            "Accept-Ranges",
-            "bytes",
+            "Access-Control-Allow-Origin",
+            "*",
             "Cache-Control",
-            "public, no-cache",
-            "Content-Length",
-            "3172",
+            "max-age=60, public",
             "Content-Type",
-            "text/html",
+            "text/plain; charset=\"utf-8\"",
             "Date",
-            "Sat, 06 Jan 2024 05:42:14 GMT",
+            "Sun, 07 Jan 2024 04:06:44 GMT",
             "Etag",
-            "\"658b31ce-c64\"",
+            "\"6596d85d-12f\"",
             "Last-Modified",
-            "Tue, 26 Dec 2023 20:04:30 GMT",
+            "Thu, 04 Jan 2024 16:10:05 GMT",
             "Server",
-            "nginx/1.25.3",
+            "gunicorn",
+            "Vary",
+            "Accept-Encoding",
+            "Content-Length",
+            "303",
             "Connection",
             "close"
         ],
@@ -506,7 +508,7 @@
             "Content-Type",
             "application/schema+json",
             "Date",
-            "Sat, 06 Jan 2024 05:42:14 GMT",
+            "Sun, 07 Jan 2024 04:06:44 GMT",
             "Referrer-Policy",
             "same-origin",
             "Server",
@@ -1000,7 +1002,7 @@
             "Content-Type",
             "application/schema+json",
             "Date",
-            "Sat, 06 Jan 2024 05:42:14 GMT",
+            "Sun, 07 Jan 2024 04:06:44 GMT",
             "Referrer-Policy",
             "same-origin",
             "Server",
