+++ test/fixtures/sdk.test.ts.json
@@ -475,7 +475,7 @@
             "Content-Type",
             "application/schema+json",
             "Date",
-            "Sat, 06 Jan 2024 20:38:40 GMT",
+            "Sun, 07 Jan 2024 04:06:43 GMT",
             "Referrer-Policy",
             "same-origin",
             "Server",
@@ -967,7 +967,7 @@
             "Content-Type",
             "application/schema+json",
             "Date",
-            "Sat, 06 Jan 2024 20:38:40 GMT",
+            "Sun, 07 Jan 2024 04:06:43 GMT",
             "Referrer-Policy",
             "same-origin",
             "Server",
